The robustness of the macro has been improved.  If GET_DLNAME is called before AC_LIBTOOL_PROG, the configure script will exit now with an appropriate error message.  This is not intended for general users, but only for developers who make changes to configure.ac.
Some comments about the last changes done by Alan:

In the GET_DLNAME macro, $LD was changed to $CC as the command when invoking
libtool --mode=link. Actually, the name of this command is irrelevant to the
actions taken by the libtool script.  However, in version 1.5 (or 1.5.0a) of
libtool, the name is used to guess a "tagged configuration", in case the
option --tag has not been specified.  Using $LD would then trigger an error.
I hope that this macro will be portable across different platforms. The
future will tell us.

svn path=/trunk/; revision=4886

+++ acinclude.m4
@@ -874,13 +874,17 @@ dnl library in the form lib$STEM.<so_ext>.<so_number>.  Set the
 dnl variable VARIABLE with the resulting value.  This macro should be used 
 dnl only after the call to AM_PROG_LIBTOOL.
 AC_DEFUN([GET_DLNAME],[
-  TMP_DIR=./tmp-cfg
-  rm -rf $TMP_DIR
-  mkdir -p $TMP_DIR
-  cd $TMP_DIR
-  ../libtool --mode=link $CC -rpath /usr/lib -version-info $2 \
-      -o lib$1.la > /dev/null
-  $3=`grep ^dlname= lib$1.la | sed "s/dlname='\(.*\)'/\1/"`
-  cd ..
-  rm -rf $TMP_DIR
+  if test -z "$LIBTOOL" -a -z "$CC" ; then
+    AC_MSG_ERROR([Dlname guessings can be done only after libtool is initialized])
+  else
+    TMP_DIR=./tmp-cfg
+    rm -rf $TMP_DIR
+    mkdir -p $TMP_DIR
+    cd $TMP_DIR
+    ../libtool --mode=link $CC -rpath /usr/lib -version-info $2 \
+        -o lib$1.la > /dev/null
+    $3=`grep ^dlname= lib$1.la | sed "s/dlname='\(.*\)'/\1/"`
+    cd ..
+    rm -rf $TMP_DIR
+  fi
 ])
